refactor: move project code to project folder
