+++ lab/admin/__init__.py
@@ -1,4 +1,4 @@
 from ..objects.admin import ObjectGroupAdmin  # noqa: F401
+from ..projects.admin import ProjectAdmin  # noqa: F401
 from .institution import InstitutionAdmin  # noqa: F401
-from .project import ProjectAdmin  # noqa: F401
 from .run import RunAdmin  # noqa: F401
