+++ lab/models/__init__.py
@@ -1,4 +1,4 @@
 from ..objects import Location, Object, ObjectGroup, Period  # noqa: F401
+from ..projects.models import BeamTimeRequest, Project  # noqa: F401
 from .participation import Institution, Participation  # noqa: F401
-from .project import BeamTimeRequest, Project  # noqa: F401
 from .run import Run  # noqa: F401
