+++ lab/projects/models.py
@@ -9,11 +9,10 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from slugify import slugify
 
+from lab.models.participation import Participation
 from lab.models.run import Run
+from lab.validators import valid_filename
 from shared.models import TimestampedModel
-
-from ..validators import valid_filename
-from .participation import Participation
 
 
 class ProjectQuerySet(models.QuerySet):
