+++ lab/projects/tests/test_change_leader_view.py
@@ -7,13 +7,14 @@ from django.core.exceptions import PermissionDenied
 from django.test.client import RequestFactory
 from django.urls import reverse
 
-from ...models import Project
-from ...views import change_leader_view
-from .. import factories
+from lab.tests import factories
+
+from .. import views
+from ..models import Project
 
 
-@mock.patch.object(change_leader_view.ChangeLeaderView, "_process_request", mock.Mock())
-@mock.patch.object(change_leader_view, "is_lab_admin", mock.Mock(return_value=True))
+@mock.patch.object(views.ChangeLeaderView, "_process_request", mock.Mock())
+@mock.patch.object(views, "is_lab_admin", mock.Mock(return_value=True))
 def test_change_leader_as_admin_is_allowed():
     request = RequestFactory().post(
         reverse("admin:lab_project_leader_participation_change", args=[1]),
@@ -22,10 +23,10 @@ def test_change_leader_as_admin_is_allowed():
     request.user = mock.Mock()
 
     # Should not raise PermissionDenied:
-    change_leader_view.ChangeLeaderView().dispatch(request, 1)
+    views.ChangeLeaderView().dispatch(request, 1)
 
 
-@mock.patch.object(change_leader_view, "is_lab_admin", mock.Mock(return_value=False))
+@mock.patch.object(views, "is_lab_admin", mock.Mock(return_value=False))
 def test_change_leader_as_non_admin_is_forbidden():
     request = RequestFactory().post(
         reverse("admin:lab_project_leader_participation_change", args=[1]),
@@ -34,7 +35,7 @@ def test_change_leader_as_non_admin_is_forbidden():
     request.user = mock.Mock()
 
     with pytest.raises(PermissionDenied):
-        change_leader_view.ChangeLeaderView().dispatch(request, 1)
+        views.ChangeLeaderView().dispatch(request, 1)
 
 
 @pytest.mark.django_db
@@ -42,7 +43,7 @@ def test_update_participation():
     project = factories.ProjectWithMultipleParticipationsFactory()
     old_leader = project.participation_set.filter(is_leader=True).get()
     new_leader = project.participation_set.filter(is_leader=False)[0]
-    change_leader_view.ChangeLeaderView._update_participation(  # pylint: disable=protected-access
+    views.ChangeLeaderView._update_participation(  # pylint: disable=protected-access
         project, new_leader.id
     )
     new_leader.refresh_from_db()
@@ -51,17 +52,17 @@ def test_update_participation():
     assert not old_leader.is_leader
 
 
-@mock.patch.object(change_leader_view.transaction, "atomic", mock.MagicMock())
-@mock.patch.object(change_leader_view.Project, "leader", mock.Mock())
-@mock.patch.object(change_leader_view.ChangeLeaderForm, "full_clean", mock.Mock())
-@mock.patch.object(change_leader_view.ChangeLeaderView, "_update_participation")
+@mock.patch.object(views.transaction, "atomic", mock.MagicMock())
+@mock.patch.object(views.Project, "leader", mock.Mock())
+@mock.patch.object(views.ChangeLeaderForm, "full_clean", mock.Mock())
+@mock.patch.object(views.ChangeLeaderView, "_update_participation")
 def test_change_leader_updates_participation(mocked_update_participation):
     project = Project(id=1)
-    form = change_leader_view.ChangeLeaderView.form_class(
+    form = views.ChangeLeaderView.form_class(
         project=project, data={"leader_participation": 42}
     )
     form.cleaned_data = {"leader_participation": mock.Mock(id=42)}
-    view = change_leader_view.ChangeLeaderView()
+    view = views.ChangeLeaderView()
     view.project = project
     view.request = RequestFactory().post(
         reverse("admin:lab_project_leader_participation_change", args=[1]),
@@ -73,20 +74,20 @@ def test_change_leader_updates_participation(mocked_update_participation):
     mocked_update_participation.assert_called_with(project, 42)
 
 
-@mock.patch.object(change_leader_view.transaction, "atomic", mock.MagicMock())
-@mock.patch.object(change_leader_view.Project, "leader", mock.Mock())
-@mock.patch.object(change_leader_view.ChangeLeaderForm, "full_clean", mock.Mock())
-@mock.patch.object(change_leader_view.ChangeLeaderView, "_update_participation")
-@mock.patch.object(change_leader_view, "TemplateResponse")
+@mock.patch.object(views.transaction, "atomic", mock.MagicMock())
+@mock.patch.object(views.Project, "leader", mock.Mock())
+@mock.patch.object(views.ChangeLeaderForm, "full_clean", mock.Mock())
+@mock.patch.object(views.ChangeLeaderView, "_update_participation")
+@mock.patch.object(views, "TemplateResponse")
 def test_change_leader_in_popup_renders_json_object(
     template_response, mocked_update_participation
 ):
     project = Project(id=1)
-    form = change_leader_view.ChangeLeaderView.form_class(
+    form = views.ChangeLeaderView.form_class(
         project=project, data={"leader_participation": 42, "_popup": 1}
     )
     form.cleaned_data = {"leader_participation": mock.Mock(id=42)}
-    view = change_leader_view.ChangeLeaderView()
+    view = views.ChangeLeaderView()
     view.project = project
     view.request = RequestFactory().post(
         reverse("admin:lab_project_leader_participation_change", args=[1]),
