--- lab/views/__init__.py
@@ -1,2 +1,0 @@
-from ..documents.views import ProjectDocumentsView  # noqa: F401
-from .change_leader_view import ChangeLeaderView  # noqa: F401
